Janitor: Refactor `filterAndSort` in `table.js` to reduce cyclomatic complexity and resolve unhandled promise rejection in service workers
